pipeline/requirements.txt
google-cloud-storage>=2.14.0
pandas>=2.0.0
requests>=2.31.0
shef-parser>=0.1.0
